*Read float instead of int for quest honor info *Fix quest poi loading from SFDB *Extract db2 files as well as dbc *Fix chat packets for AFK and DND

+++ Services/WCell.RealmServer/Quests/QuestHonor.cs
@@ -10,7 +10,7 @@ namespace WCell.RealmServer.Quests
     public class QuestHonorInfo
     {
         public int Level;
-        public int RewHonor;
+        public float RewHonor;
     }
 
     public class QuestHonorInfoConverter : ClientDBRecordConverter
@@ -21,7 +21,7 @@ namespace WCell.RealmServer.Quests
             var honorInfo = new QuestHonorInfo
             {
                 Level = GetInt32(rawData, 0) - 1,
-                RewHonor = GetInt32(rawData, 1),
+                RewHonor = GetFloat(rawData, 1),
             };
 
             QuestMgr.QuestHonorInfos[(uint)honorInfo.Level] = honorInfo;
